Sample fixes for stricter return type checking

+++ Hydrogene/Toffee/UIKit/UITableViewController/UITableViewController/RootViewController.cs
@@ -7,7 +7,7 @@ namespace CSharpiOSUITableViewControllerDemo
 	{
 
 
-	  private string[] dataArray;
+		private string[] dataArray;
 		
 		public override id init()
 		{
@@ -15,18 +15,18 @@ namespace CSharpiOSUITableViewControllerDemo
 			if (this != null)
 			{
 				title = "CSharp UITableViewController Demo";
-         this.dataArray =  new string[] {"Athens" , "Paris",	 "St. Louis" , "Athens", "London", "Stockholm", "Berlin",  "Antwerp", "Chamonix",	"St. Moritz",	"Amsterdam"	, "Los Angeles", "Lake Placid",
-                 "Garmisch-Partenkirchen", "Tokyo", "Sapporo", "Cortina d'Ampezzo", "Oslo", "Helsinki",	"Melbourne", "Stockholm", "Squaw Valley", "Rome",
-                 "Innsbruck"	, "Mexico City", "Grenoble"	,"Sapporo" ,"Munich", "Montreal", "Moscow", "Lake Placid", "Sarajevo", "Seoul", "Calgary", "Barcelona", "Albertville",
-                 "Lillehammer", "Atlanta", "Nagano", "Sydney", "Salt Lake City", "Turin", "Beijing", "Vancouver", "Sochi","Rio de Janero", "Pyeongchang"};
+				this.dataArray =  new string[] {"Athens" , "Paris",	 "St. Louis" , "Athens", "London", "Stockholm", "Berlin",  "Antwerp", "Chamonix",	"St. Moritz",	"Amsterdam"	, "Los Angeles", "Lake Placid",
+				"Garmisch-Partenkirchen", "Tokyo", "Sapporo", "Cortina d'Ampezzo", "Oslo", "Helsinki",	"Melbourne", "Stockholm", "Squaw Valley", "Rome",
+				"Innsbruck"	, "Mexico City", "Grenoble"	,"Sapporo" ,"Munich", "Montreal", "Moscow", "Lake Placid", "Sarajevo", "Seoul", "Calgary", "Barcelona", "Albertville",
+				"Lillehammer", "Atlanta", "Nagano", "Sydney", "Salt Lake City", "Turin", "Beijing", "Vancouver", "Sochi","Rio de Janero", "Pyeongchang"};
 			}
 			
 			return this;
 		}
 
 
-   #region Table view data source
-		protected int numberOfSectionsInTableView(UITableView tableView)
+		#region Table view data source
+		protected NSInteger numberOfSectionsInTableView(UITableView tableView)
 		{
 			// Return the number of sections.
 			return 1;
@@ -40,12 +40,12 @@ namespace CSharpiOSUITableViewControllerDemo
 		protected UITableViewCell tableView(UITableView tableView) cellForRowAtIndexPath(NSIndexPath indexPath)
 		{
 			UITableViewCell cell = tableView.dequeueReusableCellWithIdentifier("SimpleTableCell");
-      if (cell == null)
-         cell = new UITableViewCell withStyle(UITableViewCellStyle.UITableViewCellStyleDefault) reuseIdentifier("SimpleTableCell");
+			if (cell == null)
+				cell = new UITableViewCell withStyle(UITableViewCellStyle.UITableViewCellStyleDefault) reuseIdentifier("SimpleTableCell");
 
-       cell.textLabel.text = this.dataArray[indexPath.row];
+			cell.textLabel.text = this.dataArray[indexPath.row];
 
-       return cell; 
+			return cell; 
 		}
 	
 		#endregion
@@ -53,9 +53,9 @@ namespace CSharpiOSUITableViewControllerDemo
 		#region Table view delegate
 		protected void tableView(UITableView tableView) didSelectRowAtIndexPath(NSIndexPath indexPath)
 		{
-         string text = tableView.cellForRowAtIndexPath(indexPath).textLabel.text;
-         UIAlertView message = new UIAlertView withTitle("City Selected") message(text) @delegate(null) cancelButtonTitle("OK") otherButtonTitles(null);
-         message.show();
+			string text = tableView.cellForRowAtIndexPath(indexPath).textLabel.text;
+			UIAlertView message = new UIAlertView withTitle("City Selected") message(text) @delegate(null) cancelButtonTitle("OK") otherButtonTitles(null);
+			message.show();
 		}
 
 		#endregion
